# RacingGuru/internal/controller/handlers/assets/swagger/swagger.yaml
basePath: /
definitions:
...
  handlers.ErrorResponse:
    properties:
      error:
        properties:
          code:
            type: string
          message:
            type: string
        type: object
    type: object
...
  models.ConditionOperator:
    enum:
    - gt
    - lt
    - gte
    - lte
    - eq
    type: string
    x-enum-varnames:
    - ConditionOperatorGT
    - ConditionOperatorLT
    - ConditionOperatorGTE
    - ConditionOperatorLTE
    - ConditionOperatorEQ
  models.Driver:
    properties:
      birthday:
        type: string
      id:
        type: string
      name:
        type: string
      nationality:
        type: string
    type: object
...
  models.MatrixDrivers:
    properties:
      condition_specs:
        items:
          items:
            $ref: '#/definitions/models.SudokuCondition'
          type: array
        type: array
      field:
        items:
          items:
            items:
              $ref: '#/definitions/models.Driver'
            type: array
          type: array
        type: array
    type: object
  models.MatrixTeams:
    properties:
      condition_specs:
        items:
          items:
            $ref: '#/definitions/models.SudokuCondition'
          type: array
        type: array
      field:
        items:
          items:
            items:
              $ref: '#/definitions/models.Team'
            type: array
          type: array
        type: array
    type: object
...
  models.SudokuCondition:
    properties:
      field:
        type: string
      label:
        type: string
      op:
        $ref: '#/definitions/models.ConditionOperator'
      value:
        type: integer
    type: object
  models.Team:
    properties:
      country:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
info:
  contact: {}
  description: HTTP API for authentication and racing statistics.
  title: RacingGuru API
  version: 0.1.0
paths:
...
  /sudoku/drivers:
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MatrixDrivers'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get driver sudoku matrix
      tags:
      - sudoku
  /sudoku/teams:
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MatrixTeams'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get team sudoku matrix
      tags:
      - sudoku
...
swagger: "2.0"